Finished Loan Shark
App.html: Adjusted the UI to have it display the table and results of the calculations better, this took me some time and pretty much relearned many aspects of Bootstrap 5 that I have forgotten.

Code.html: mainly just maintenance tasks, added my loan shark logo to the title images and navbar. And copied the JS code from site.js into the prism.js tags to display. Had to adjust the bootstrap columns into one so that most of the JS code would display correctly in the black box.  And finally added the descriptions to the best of my ability, discussing the details of the JS functions.

Site.js: did some minor edits to the spacing.